refactor : 기존 테스트 코드 수정(H2 DB 테스트 실행 시점에 Drop 후 다시 Create
