+++ src/main/resources/application-test.yml
@@ -8,3 +8,6 @@ spring:
     init:
       mode: always
       schema-locations: file:./sql/table.sql
+  jpa:
+    hibernate:
+      ddl-auto: none
