+++ src/test/java/com/ticket/concert/ConcertApplicationTests.java
@@ -2,7 +2,9 @@ package com.ticket.concert;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 
+@ActiveProfiles("test")
 @SpringBootTest
 class ConcertApplicationTests {
